Break loader dependency on wasmtime
