+++ crates/outbound-http/Cargo.toml
@@ -11,9 +11,14 @@ doctest = false
 anyhow  = "1.0"
 http = "0.2"
 reqwest = { version = "0.11", features = ["gzip"] }
-spin-app = { path = "../app" }
-spin-core = { path = "../core" }
-spin-world = { path = "../world" }
+spin-app = { path = "../app", optional = true }
+spin-core = { path = "../core", optional = true }
+spin-locked-app = { path = "../locked-app" }
+spin-world = { path = "../world", optional = true }
 terminal = { path = "../terminal" }
 tracing = { workspace = true }
 url = "2.2.1"
+
+[features]
+default = ["runtime"]
+runtime = ["dep:spin-app", "dep:spin-core", "dep:spin-world"]
