+++ crates/outbound-http/src/host_component.rs
@@ -4,7 +4,7 @@ use spin_app::DynamicHostComponent;
 use spin_core::{Data, HostComponent, Linker};
 use spin_world::v1::http;
 
-use crate::{allowed_http_hosts::parse_allowed_http_hosts, OutboundHttp};
+use crate::{allowed_http_hosts::parse_allowed_http_hosts, host_impl::OutboundHttp};
 
 pub struct OutboundHttpComponent;
